+++ src/app/app.module.ts
@@ -4,16 +4,14 @@ import { NgModule } from '@angular/core';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 
-import { CartService } from './modules/cart/services/cart.service';
 import { CartModule } from './modules/cart/cart.module';
 import { OrdersModule } from './modules/orders/orders.module';
 import { ProductModule } from './modules/product/product.module';
-import { ProductService } from './modules/product/services/products.service';
 import { FirstModule } from './modules/first/first.module';
 import { Router } from '@angular/router';
-import { LocalStorageService } from './core/services/local-storage.service';
-import { OrderService } from './modules/orders/order.component/services/order.service';
-import { AdminModule } from './modules/admin/admin.module';
+import { HttpClientModule } from '@angular/common/http';
+import { SharedModule } from './shared/shared.module';
+import { httpInterceptorProviders } from './core/interceptors';
 
 @NgModule({
   declarations: [
@@ -21,17 +19,16 @@ import { AdminModule } from './modules/admin/admin.module';
   ],
   imports: [
     BrowserModule,
+    HttpClientModule,
+    SharedModule,
     CartModule,
     OrdersModule,
-    AdminModule,
     ProductModule,
     FirstModule,
     AppRoutingModule,
   ],
   providers: [
-    ProductService,
-    CartService,
-    OrderService
+    httpInterceptorProviders
   ],
   bootstrap: [AppComponent]
 })