+++ src/app/core/guards/auth.guard.ts
@@ -11,7 +11,7 @@ import { AuthService } from '../services/auth.service';
 })
 export class AuthGuard implements CanActivate {
     constructor(private authService: AuthService,
-        private router: Router) {
+                private router: Router) {
     }
 
     canActivate(