+++ src/app/core/models/app.settings.model.ts
@@ -1,0 +1,7 @@
+
+export class AppSettingsModel {
+    constructor(public appName: string,
+        public auther: string,
+        public company: string) {
+    }
+}