+++ src/app/modules/admin/guards/resolve.guard.ts
@@ -1,7 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Resolve, ActivatedRouteSnapshot } from '@angular/router';
 
-import { Observable, of } from 'rxjs';
 import { ProductModel } from '../../product/models/product.model';
 import { ProductService } from '../../product/services/products.service';
 
@@ -14,15 +13,13 @@ export class ProductResolveGuard implements Resolve<ProductModel> {
         private productsService: ProductService
     ) { }
 
-    resolve(route: ActivatedRouteSnapshot): Observable<ProductModel | null> {
+    resolve(route: ActivatedRouteSnapshot): Promise<ProductModel | null> {
 
         if (!route.paramMap.has('productId')) {
-            return of(new ProductModel(null, null, null, null, 'assets/images/iphoneXs.jpg', null));
+            return Promise.resolve(new ProductModel(null, null, null, null, 'assets/images/iphoneXs.jpg', null));
         }
 
         const id = +route.paramMap.get('productId');
-        const product = this.productsService.getProductById(id);
-
-        return of(product);
+        return this.productsService.getProductById(id);
     }
 }