+++ src/app/modules/orders/services/order.service.ts
@@ -1,7 +1,7 @@
 
 import { Injectable } from '@angular/core';
 import { LocalStorageService } from 'src/app/core/services/local-storage.service';
-import { OrderModel } from '../../models/order.model';
+import { OrderModel } from '../models/order.model';
 
 @Injectable()
 export class OrderService {